refactor: match folder structure of original source
