+++ types/nodemailer/lib/base64/index.d.ts
