+++ types/nodemailer/lib/ses-transport/index.d.ts
@@ -2,13 +2,13 @@
 
 import { EventEmitter } from 'events';
 
-import { Transport, TransportOptions } from '..';
+import { Transport, TransportOptions } from '../..';
 
-import * as shared from './shared';
+import * as shared from '../shared';
 
-import Mail = require('./mailer');
-import MailMessage = require('./mailer/mail-message');
-import MimeNode = require('./mime-node');
+import Mail = require('../mailer');
+import MailMessage = require('../mailer/mail-message');
+import MimeNode = require('../mime-node');
 
 declare namespace SESTransport {
     interface MailOptions extends Mail.Options {
