+++ types/nodemailer/lib/stream-transport/index.d.ts
@@ -3,13 +3,13 @@
 import { EventEmitter } from 'events';
 import { Readable } from 'stream';
 
-import { Transport, TransportOptions } from '..';
+import { Transport, TransportOptions } from '../..';
 
-import * as shared from './shared';
+import * as shared from '../shared';
 
-import Mail = require('./mailer');
-import MailMessage = require('./mailer/mail-message');
-import MimeNode = require('./mime-node');
+import Mail = require('../mailer');
+import MailMessage = require('../mailer/mail-message');
+import MimeNode = require('../mime-node');
 
 declare namespace StreamTransport {
     type MailOptions = Mail.Options;
