+++ types/nodemailer/lib/well-known/index.d.ts
@@ -1,4 +1,4 @@
-import SMTPConnection = require('./smtp-connection');
+import SMTPConnection = require('../smtp-connection');
 
 /** Resolves SMTP config for given key. Key can be a name (like 'Gmail'), alias (like 'Google Mail') or an email address (like 'test@googlemail.com'). */
 declare function wellKnown(key: string): SMTPConnection.Options | false;
